增加1个服务，1个模块，1个用户控件 1. 添加联系人 (Contact) 模块 2. 增加名片用户控件。 3. 增加Dialog服务。

+++ GenmCloud.Core/Data/VO/ChatObjectVO.cs
@@ -11,7 +11,7 @@ namespace GenmCloud.Core.Data.VO
 
     public class ChatObjectVO : BindableBase
     {
-        public long Id { get; set; }
+        public uint Id { get; set; }
         public string Name { get; set; }
         public string LastMsg { get; set; }
         public long LastMsgTimestamp { get; set; }
